Got Captcha working. dicided to use it on the front-end. Now seting up assets folder and wrapping up version 1

+++ include/live-update.php
@@ -2,7 +2,7 @@
 
 include_once 'dbh.php';
 
-$sql = 'SELECT user_name, u.user_id, w.user_id, wrk_name, w.type_id, t.type_id, type_name FROM users u RIGHT JOIN workout_desc w ON u.user_id = w.user_id LEFT JOIN workout_type t ON w.type_id = t.type_id ORDER BY wrk_id DESC LIMIT 7';
+$sql = 'SELECT user_name, u.user_id, w.user_id, wrk_name, w.type_id, t.type_id, type_name FROM users u RIGHT JOIN workout_desc w ON u.user_id = w.user_id LEFT JOIN workout_type t ON w.type_id = t.type_id ORDER BY wrk_id DESC LIMIT 30';
 $result = mysqli_query($conn, $sql);
 if (mysqli_num_rows($result) > 0) {
   while ($row = mysqli_fetch_assoc($result)) {
@@ -10,9 +10,17 @@ if (mysqli_num_rows($result) > 0) {
 
     $fileExt = 'jpg' or 'jpeg' or 'png' or 'gif';
 
+  if (file_exists('uploads/'.$row['user_name'].'_profile.jpg')) {
       echo '<div class="online">';
       echo '<div class="user-photo"><a href="#userPic"><img src="uploads/'.$row['user_name'].'_profile.'.$fileExt.'" alt="'.$row['user_name'].'"></a></div>';
       echo '<p class="name">'.$row['user_name'].' created a  new workout '.$row['wrk_name'].'</p>';
       echo '</div>';
+  }
+  else {
+    echo '<div class="online">';
+      echo '<div class="user-photo"><a href="#userPic"><img src="uploads/profiledefault.png" alt="Default profile image for Your Progress vs Mine"></a></div>';
+      echo '<p class="name">'.$row['user_name'].' created a  new workout '.$row['wrk_name'].'</p>';
+      echo '</div>';
+  }
   }
 }
